Extract citeproc helper functions to extra crate
