+++ crates/bibtex-utils/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "citeproc"
+name = "bibtex-utils"
 version = "0.0.0"
 license.workspace = true
 authors.workspace = true
@@ -7,15 +7,13 @@ edition.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-bibtex-utils = { path = "../bibtex-utils" }
-isocountry = "0.3.2"
+chrono = { version = "0.4.26", default-features = false, features = ["std"] }
+human_name = "2.0.2"
 itertools = "0.11.0"
 rowan = "0.15.11"
 rustc-hash = "1.1.0"
 syntax = { path = "../syntax" }
-titlecase = "2.2.1"
 unicode-normalization = "0.1.22"
-url = "=2.3.1"
 
 [lib]
 doctest = false
