+++ crates/hover/Cargo.toml
@@ -8,6 +8,7 @@ rust-version.workspace = true
 
 [dependencies]
 base-db = { path = "../base-db" }
+bibtex-utils = { path = "../bibtex-utils" }
 citeproc = { path = "../citeproc" }
 completion-data = { path = "../completion-data" }
 rowan = "0.15.11"
