+++ crates/syntax/src/lib.rs
@@ -20,16 +20,3 @@ pub struct BuildError {
 pub struct BuildLog {
     pub errors: Vec<BuildError>,
 }
-
-#[macro_export]
-macro_rules! match_ast {
-    (match $node:ident { $($tt:tt)* }) => { $crate::match_ast!(match ($node) { $($tt)* }) };
-
-    (match ($node:expr) {
-        $( $( $path:ident )::+ ($it:pat) => $res:expr, )*
-        _ => $catch_all:expr $(,)?
-    }) => {{
-        $( if let Some($it) = $($path::)+cast($node.clone()) { $res } else )*
-        { $catch_all }
-    }};
-}
